# Conflicts:
#	src/main/java/group/ship/blackshipstore/entity/Order.java
#	src/main/java/group/ship/blackshipstore/entity/Status.java
#	src/main/java/group/ship/blackshipstore/repositories/ArticlesOrdersRepository.java
#	src/main/java/group/ship/blackshipstore/repositories/ArticlesRepository.java
#	src/main/java/group/ship/blackshipstore/repositories/OrderRepository.java
#	src/main/java/group/ship/blackshipstore/repositories/PirateRepository.java
#	src/main/java/group/ship/blackshipstore/repositories/Repository.java
#	src/main/java/group/ship/blackshipstore/sevices/ItemService.java
#	src/main/java/group/ship/blackshipstore/sevices/OrderService.java
#	src/main/resources/db/migration/V1__initDB.sql
#	src/main/resources/db/migration/V2_initDB.sql

+++ src/main/java/group/ship/blackshipstore/controller/OrderController.java
@@ -25,13 +25,13 @@ public class OrderController {
     }
     @GetMapping("/{id}")
     @Operation(summary = "Get all Pirate Orders", description = "Return list of all Orders by Pirate id")
-    public List<Order> getAllOrdersByPirateId(@PathVariable Integer id) {
+    public List<Order> getAllOrdersByPirateId(@PathVariable Long id) {
         return orderService.getAllOrdersByPirateId(id);
     }
 
     @GetMapping("/unique/{id}")
     @Operation(summary = "Get last Pirate Order", description = "Return last Order by Pirate id")
-    public Order getLastOrderByPirateId(@PathVariable Integer id) {
+    public Order getLastOrderByPirateId(@PathVariable Long id) {
         return orderService.getLastOrderByPirateId(id);
     }
